Suggestions: Update ownership of core files and improve some types
