+++ public/app/features/panel/suggestions/getAllSuggestions.ts
