+++ public/app/plugins/panel/radialbar/suggestions.ts
@@ -1,5 +1,6 @@
 import { VisualizationSuggestionsBuilder } from '@grafana/data';
-import { FieldColorModeId } from '@grafana/schema/dist/esm/index.gen';
+import { FieldColorModeId } from '@grafana/schema';
+import { GraphFieldConfig } from '@grafana/ui';
 import { SuggestionName } from 'app/types/suggestions';
 
 import { Options } from './panelcfg.gen';
@@ -17,7 +18,7 @@ export class GaugeSuggestionsSupplier {
       return;
     }
 
-    const list = builder.getListAppender<Options, {}>({
+    const list = builder.getListAppender<Options, GraphFieldConfig>({
       name: SuggestionName.Gauge,
       pluginId: 'gauge',
       options: {},
@@ -27,8 +28,8 @@ export class GaugeSuggestionsSupplier {
       },
       cardOptions: {
         previewModifier: (s) => {
-          if (s.options!.reduceOptions.values) {
-            s.options!.reduceOptions.limit = 2;
+          if (s.options?.reduceOptions?.values) {
+            s.options.reduceOptions.limit = 2;
           }
         },
       },
